gnomebase.bbclass: Use meson as default buildsystem
The vast majority of gnome recipes uses meson. Set it as default
and override the few recipes that still use autotools.

This way we can remove a lot of lines in meta-oe and more important
it would not be needed to explicitly set GNOMEBASEBUILDCLASS = "meson" for newly
created gnome recipe anymore.

Signed-off-by: Markus Volk <f_l_k@t-online.de>
Signed-off-by: Alexandre Belloni <alexandre.belloni@bootlin.com>