gpgme: disable python support (until upstream fixes 3.12 compatibility)
The only consumer in oe-core was dnf, and latest dnf releases no longer
use it. There are no consumers in meta-oe.

gpgme's python support is generally a pain to support, and lags
behind upstream python versions. It was enabled in

commit fc6c81da79897c95a9c81960a1dee49517f2562a
Author: Mark Hatle <mark.hatle@windriver.com>
Date:   Tue Mar 7 12:54:33 2017 -0600

    gpgme: Fix issue building for the target

which wasn't prompted by any consumer requiring python,
but rather to correct build errors in absence of specific
configuration passed in.

Signed-off-by: Alexander Kanavin <alex@linutronix.de>
Signed-off-by: Alexandre Belloni <alexandre.belloni@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>