systemd-bootchart: upgrade from 234 to 235
0001-architecture-Recognise-RISCV-32-RISCV-64.patch is dropped because
it has some problem. It's a patch from systemd repo, yet it's marked
as 'Backport' for systemd-bootchart. systemd and systemd-bootchart are
two different projects. Now in 235, RISCV64 has been officially supported,
but RISCV32 is not. So drop this patch. People who want RISCV32 support
need to work with systemd-bootchart upstream first.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>
Signed-off-by: Alexandre Belloni <alexandre.belloni@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>