gawk: Add coreutils to rdeps for ptests package
newer gawk 5.3.0+ added a new testcase test/cmdlinefsbacknl2.sh which
use od utility with -An option which is not available in busybox
provided od applet. Replacing -An is not possible because default output
for od is octal and the test wants to use none for radix thats why it
uses -An.

Fixes test/cmdlinefsbacknl2.sh failures in ptests

which fails like below

   Usage: od [-aBbcDdeFfHhIiLlOovXx] [FILE]
    od: unrecognized option: A
    BusyBox v1.36.1 () multi-call binary.

Sadly, this prolongs depchain if ptest distro feature is used but there
is no other option if this test has to execute. Perhaps its best to
report this to upstream gawk and request to fix the testcase to not
hard depend on coreutils/od utility

Signed-off-by: Khem Raj <raj.khem@gmail.com>
Signed-off-by: Alexandre Belloni <alexandre.belloni@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>