cairo: upgrade to 1.18.0
The changelog is large[1] but:

Added:
- Type 3 colour and COLRv2 fonts
- Tee surfice automatically enabled
- Meson build system

Removed:
- The cairo-sphinx tool
- The XML, GL, GLES, Qt5, BeOS, OS/2, DirectFB, DRM, Cogl, and OpenVL
  backends
- Autotools build system

These CVE patches have been merged upstream:
  CVE-2018-19876.patch: https://gitlab.freedesktop.org/cairo/cairo/-/commit/90e85c2493fdfa3551f202ff10282463f1e36645
  CVE-2019-6461.patch: https://gitlab.freedesktop.org/cairo/cairo/-/commit/09643ee1abdd5daacebfcb564448f29be9a79bac
  CVE-2019-6462.patch: https://gitlab.freedesktop.org/cairo/cairo/-/commit/bbeaf08190d3006a80b80a77724801cd477a37b8
  CVE-2020-35492.patch: https://gitlab.freedesktop.org/cairo/cairo/-/commit/c986a7310bb06582b7d8a566d5f007ba4e5e75bf

Some niche options have been removed in the migration to Meson, so the
recipe can be cleaned up.

Put libcairo-fdr ("flight data recorder") in the perf-utils package
instead of deleting it.

There is an executable test for IPC behaviour that needs to be set via a
cross file if qemu-user isn't available, thanks to Alex Kanavin for this
portion.

[1] https://www.cairographics.org/news/cairo-1.18.0/

Signed-off-by: Ross Burton <ross.burton@arm.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>