glib-2.0: ensure GI_DATA_ENABLED is set
With python 3.12 some of the glib ptests started failing. Inspection
revealed that they fail because all tests in some glib's test suites are skipped;
python 3.12's unittest module deems "no tests in a test suite were executed"
a failure (and I tend to agree).

Why are all the tests skipped? Because python dbus module is missing from
the image, and it's missing because it's pulled in by dbusmock, which
in turn is missing because it is pulled in by glib recipe subject to
GI_DATA_ENABLED, and that variable is not global but defined in a g-i-data
class that glib recipe did not include.

So needed dependencies were simply always excluded regardless of g-i settings,
until now.

Signed-off-by: Alexander Kanavin <alex@linutronix.de>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>