libtool: Update cleanup sysroot handling patch
The sysroot path cleanup change was reworked after discussion/review upstream. It
appeared there were two different changes in one of our patches so separate out
the second one and mark as inappropriate as it is unlikely upstream would take
such a change in defaults. We need further investigation to decide if OE really
wants/needs this going forward.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>