libunwind: clean up configuration
Add a PACKAGECONFIG for the support for zlib-compressed debuginfo.

There's no need to explicitly enable reading of .debug_frame segments as
this is enabled automatically for arm/aarch64.

Enable static globally, there's not much to gain by making this
musl-specific.

Don't disable the tests on musl as they verify that libunwind can link,
which is important with external libucontext libraries.

No need to disable documentation on musl, as it's already disabled out
of the box.

Signed-off-by: Ross Burton <ross.burton@arm.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>