linux-firmware: fix mediatek MT76x empty license package
Installing the linux-firmware-dev package fails because of the following problem.
For each mediatek MT76x firmware a separate license package was declared.
In all these license packages the same file was referenced as the license file.
This meant that if several of these packages were used at the same time,
the license packages would be empty.

This commit defines a common license package for the mediatek MT76x packages,
which is then referenced by all of them.

Fixes: dfb7d2c426b4 ("linux-firmware: create separate packages")
Signed-off-by: Julian Haller <julian.haller@philips.com>
Signed-off-by: Alexandre Belloni <alexandre.belloni@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>