u-boot: add missing dependency on pyelftools-native
When TF-A is necessary in U-Boot binary, binman requires elftools to be
installed to be able to generate that U-boot ITB image.

TF-A is necessary for at least all Aarch64 machines, so that is a
non-negligible amount of boards that have this requirement.

As a side note, Rockchip-based machines didn't need this until commit
12c3e948eeab ("rockchip: Drop the FIT generator script") (v2023.04-rc1).

This is already in meta-rockchip, c.f.
https://git.yoctoproject.org/meta-rockchip/commit/recipes-bsp/u-boot?id=6127d169acf239a53df989f34a6b825fa182cc0c
but I feel like this makes more sense to be present in OE-Core.

Cc: Quentin Schulz <foss+yocto@0leil.net>
Signed-off-by: Quentin Schulz <quentin.schulz@theobroma-systems.com>
Signed-off-by: Alexandre Belloni <alexandre.belloni@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>