selftest/SStateCacheManagement: do not manipulate ERROR_QA
This is no longer necessary as selftest enforces buildhistory disabling nowadays.

It is also a significant contributor to slowness of the tests:
they work by setting up private sstate and retaining autobuilder sstate
as a read-only mirror, and if the needed objects aren't in that
mirror (as particularly seen on arm hosts), then the whole build
stars from scratch - and then is thrown away.

Signed-off-by: Alexander Kanavin <alex@linutronix.de>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>